Apple scripts: updates and new Config script
The "BittyTax Accounting Tool" has been updated to prompt the user for any extra command line arguments. If the PDF report is successfully created the user will be given the option to open the file or show the file in the Finder.

The "BittyTax Conversion Tool" has been updated to prompt the user for any extra command line arguments. If the conversion is successful the user will be given the option to open the file or show the file in the Finder.

The "BittyTax Audit Tool" has been removed, instead use the "BittyTax Accounting Tool" and enter "--audit" when prompted for command line arguments.

A new "BittyTax Config" script has been added, this will open the bittytax.conf in TextEdit.

The "BittyTax Config", "BittyTax Install (Latest Release)" and "BittyTax Update (Unreleased)" scripts, all prompt the user if they would like to run the script or not.