Accounting tool: Excel tax report now shows origins of tax calculations
The Excel tax report now contains a worksheet containing all the buys and sells which are split from the original transaction records.

Internal hyperlinks are provided to highlight when buy(s)/sell is using for the various tax calculations.

There are also external hyperlinks from the buys/sells to the transaction records which they originated to provide full traceability.