Coinbase parser: fix missing warning for advanced trades which are crypto-to-crypto
For an Advanced Trade Buy/Sell, if the quote asset of the trading pair in the "Notes" field (i.e. BTC-USDT) does not match the local currency (i.e. GBP), a warning should be produced. This indicates that the quote asset (i.e. in this case USDT) will not balance, since the Coinbase file contains the amount/fee already converted to the local currency.

This warning can also apply to normal Buy/Sell trades, where the currency specified in the "Notes" field, does not match the local currency, (i.e. EUR instead of GBP).

I have also corrected/improved the regex to match an asset symbol of more than 3 characters (i.e. USDT), and the "for" amount can optionally be prefixed with a currency symbol.

This change also cointains another fix, for data files which are not already in the local currency. The subtotal and fees were not being correclty converted into the local currency.