CoinCorner parser: added new export format
Updated the CoinCorner parser for new export format. You will need to process both the BTC and GBP export files at the same time.

There appears to be bug in the GBP report, after a certain date all transactions are classified with a type of "FPS". The parser will do its best to identify what the transaction actually is, i.e. trade or bank deposit/withdrawal.