DeFi Taxes parser: multiple addresses in filename
If the DeFi Taxes parser cannot identify your own wallet address you can include it in the filename. The parser has been updated to only require the first 10 characters of the address in the filename, this gives more room when multiple addresses are required.

Also, added additional handling for "borrow" and "repay".