BloFin parser: added funding fee export
Updated parser for funding fees export file.

There is no way to precisely correlate which funding fees relate to which futures trades. Instead funding fees with matching symbols are accumulated, until there have been no funding fees within a 24hr period, assuming the position must have been closed.

Also, the futures parser has been updated so if a position is closed which does not exist. Instead of failing with an exception, a warning is produced and a dummy position created to allow it be closed without error.