Accounting tool: split tax rates for 2024/25
For a split tax year, the capital gains summary section will show your gains:

1. Before the split date (i.e. 30th Oct 2024)
2. On or after the split date

The CGT estimate section will give 2 estimates for each of the gains above.

The "Annual Exempt Amount" tax-free allowance (which is £3,000 in 2024/25) is applied in the most beneficial way, i.e. for 2024/25 it will be applied to the gains on or after the split date which have the higher rates first. Any remaining allowance will then be used for gains at the lower rates.

I have also added the new tax rates for 2025/2026 which at the time of writing apply to the whole tax year.