Updated mapping for some of the new operations to transaction types
Updated "max-module-lines" in pylint config to prevent error, and to keep parser as one module, this makes tracking of the changes easier.

Changes:

"Asset - Transfer" to be an Airdrop
"Binance Card Spending" credit to be a Fee-Rebate
"Crypto Box" to be Income, since you are performing tasks to get the tokens, see https://www.binance.com/en/square/post/597148.