Accounting tool: added "Per-Wallet" cost tracking
New config parameters added:

1. cost_tracking_method

This parameter controls the cost tracking method to be used.

0 = Universal - legacy method (default)
1 = Per-Wallet
2 = Per-Wallet (from 1 Jan 2025)

Universal is the default (if not configured). This legacy method keeps your asset costs in a single pool, ignoring which wallet it came from.

Per-Wallet tracks your costs per individual wallet. This requires all wallet transfers (Withdrawal followed by a Deposit) to be defined correctly. BittyTax assumes transfers are happening in time sequence. If transfers are not correct it could result in a cost basis zero being used.

Option 2. Is a combination of both methods, for IRS compliance. For pre-2025, "Universal" tracking is used, and then from 1st Jan 2025 it's "Per-Wallet".

2. universal_ordering

This parameter controls how transactions are ordered for the "Universal" cost tracking method.

0 = TID (default)
1 = Date - legacy ordering

Each transaction is allocated a unique Transaction ID (TID) to keep them time-ordered. If transactions have the exact same timestamp, then file ordering is maintained.

For "Per-Wallet" cost tracking, TID ordering is mandatory. This ensures that transfers between wallets happen in the correct sequence.

Date ordering is the legacy method. When transactions are ordered by date, the strict sequencing of transactions is not maintained. When an asset is bought/sold on the same day, the purchase could be happening after the disposal.

Note: the cost basis method (FIFO, LIFO, etc) are applied universally. It is not possible to change the cost basis method for an individual wallet.